Develop

+++ src/components/addRoom/addImages/ImagesList.js
@@ -10,8 +10,15 @@ import deleteFile from '../../../firebase/deleteFile'
 import {v4 as uuidv4} from 'uuid'
 
 const ImagesList = () => {
+  const {
+    state: {images, currentUser, updatedRoom},
+    dispatch,
+  } = useValue()
+
   const handleDelete = async (image) => {
     dispatch({type: 'DELETE_IMAGE', payload: image})
+    if (updatedRoom)
+      return dispatch({type: 'UPDATE_DELETED_IMAGES', payload: [image]})
     const imageName = image?.split(`${currentUser?.id}%2F`)[1].split('?')[0]
     try {
       await deleteFile(`rooms/${currentUser?.id}/${imageName}`)
@@ -19,11 +26,6 @@ const ImagesList = () => {
       console.log(error)
     }
   }
-
-  const {
-    state: {images, currentUser},
-    dispatch,
-  } = useValue()
 
   return (
     <ImageList
